A lot of changes in different assignments

+++ Week6_REST/Rest_JAXRS_Client_AJAX_Day2/src/main/java/interfaces/IPersonFacade.java
@@ -16,8 +16,8 @@ import javax.persistence.EntityManagerFactory;
 public interface IPersonFacade {
     public void addEntityManagerFactory(EntityManagerFactory emf);
     public Person addPerson(Person p);
-    public Person deletePerson(int id);
-    public Person getPerson(int id);
+    public Person deletePerson(Long id);
+    public Person getPerson(Long id);
     public List<Person> getPersons();
     public Person editPerson(Person p);
 }
